+++ proxy/changelog.md
@@ -1,3 +1,4 @@
 ### Changes in 2.1.7
 - Fixed `/vreload` not reloading language changes.
 - Improved performance using epoll when available.
+- Fixed `java.util.ConcurrentModificationException` in ProxyChannelHandler.